Fixed validation warning with swapchain present index
Removing the present semaphore was the wrong fix as the semaphore is needed to sync the present and render thread. The issue was that the semaphore value was wrongly set to the previous frame instead of the current one that finished presenting because the value always has to be one ahead, e.g. setting the semaphore to 1 if frame 0 finished